# 🔧⚡ Fix: Memory Tools, LLM Performance & Error Handling