… docs/fixes/memory_tool_llm_decision_fix.md
… docs/fixes/turso_session_restoration_error_handling.md
… docs/optimizations/llm_pre_initialization.md